+++ src/jwk.c
@@ -13,7 +13,6 @@
 #include <cjose/base64.h>
 #include <cjose/util.h>
 
-#include <assert.h>
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
@@ -190,8 +189,9 @@ bool cjose_jwk_release(cjose_jwk_t *jwk)
         jwk->kid = NULL;
 
         // assumes freefunc is set
-        assert(NULL != jwk->fns->free_func);
-        jwk->fns->free_func(jwk);
+        if (NULL != jwk->fns->free_func) {
+            jwk->fns->free_func(jwk);
+        }
         jwk = NULL;
     }